if applied, will REFACTOR: overall project structure, friends component into friends folder and groups in groups folder
