+++ expense-splitter/src/components/Groups/GroupExpenseBar.jsx
