+++ expense-splitter/src/components/Groups/GroupName.jsx
