+++ expense-splitter/src/components/Groups/GroupSmallExpenseCard.jsx
